+++ learners/conf/db_models.py
@@ -1,0 +1,40 @@
+# sourcery skip: avoid-builtin-shadow
+from learners.database import db
+from sqlalchemy import func
+
+
+class User(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    username = db.Column(db.String(20), unique=True, nullable=False)
+    executions = db.relationship("Execution", backref="user", lazy=True)
+
+
+class Execution(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    type = db.Column(db.String(120), nullable=False)
+    script = db.Column(db.String(120), nullable=True)
+    execution_timestamp = db.Column(db.DateTime, nullable=False, default=func.current_timestamp())
+    response_timestamp = db.Column(db.DateTime, nullable=True)
+    response_content = db.Column(db.Text, nullable=True)
+    form_data = db.Column(db.String(), nullable=True)
+    msg = db.Column(db.String(240), nullable=True)
+    uuid = db.Column(db.String(120), unique=True, nullable=True)
+    completed = db.Column(db.Integer, nullable=False, default=0)
+    connection_failed = db.Column(db.Integer, nullable=False, default=0)
+    user_id = db.Column(db.Integer, db.ForeignKey("user.id"), nullable=False)
+    exercise_id = db.Column(db.Integer, db.ForeignKey("exercise.id"), nullable=False)
+
+
+class Exercise(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    type = db.Column(db.String(120), nullable=False)
+    name = db.Column(db.String(120), unique=True, nullable=False)
+    pretty_name = db.Column(db.String(120), unique=True, nullable=False)
+    weight = db.Column(db.String(120), nullable=False)
+    executions = db.relationship("Execution", backref="exercise", lazy=True)
+
+
+class TokenBlocklist(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    jti = db.Column(db.String(36), nullable=False)
+    created_at = db.Column(db.DateTime, nullable=False)
